Refactor index.js naming convention
